Added Orientation to Shred Class
The shred class was missing a parameter to calculate the orientation
An enum Orientation with Regular and Reversed options was added.
As a result all previous measures, getters and plotters had to be
converted to support two parameters, Orientation and Direction.

Lastly, Utility.Reverse and a test method was added for this.