Implemented Query Interface(NotTested)
Given a Data Object, one can query a Shred for whether it is a feasible
match.  This works in the following way.

We get the Shreds in question.
We get their roots and ensure they don't match
We get then take the query, converted it to the orientation
Check if it matches, and build a query that does match.
Finally we ensure that the edges in question are available for
the given query direction