Restructured Cluster Constructor
Removed the Root Validation as it is done in the Match function.

THis is a costly step and doesn't need to be repeated if we can avoid
it.

Added a private Data _data member to store data that can be used to
get the offset of the match between 2 shreds.