Implement JigsawTest Cluster Visualizer
Cluster Visualizer is a Debugging Tool that is Used to show the
composition of the shred.  It can be extended to show images etc.

To export the necessary data, the match info on the cluster, was exposed
through the INode.MatchData Method.

Added the necessary files for visualization to the visualizer/* folder.
This folder is copied when altered to the Test projects output directory

The readme was updated to document a potential hiccup from VS and Chrome.
The gitignore was updated to add the data.js