Correct Shred Constructor Bug
There was a mistake in the shred Constructor, in which the data was later
being indexed correctly, but the constructor put it into the wrong position.

This is now corrected. Visualizer.js was amended to display node images correctly
now.

There is still some problem with the mirroring, it appears that visualizers' nodes aren't flipping.
This is kind of strange.

This breaks some Reconstruction Tests (NaiveKruskallPrimitive)